VectorModel

Missing docstring.

Missing docstring for IdFreq. Check Documenter's build log for details.

TextSearch.VectorModel — Type

VectorModel(global_weighting::GlobalWeighting, local_weighting::LocalWeighting, corpus; mindocs=1)

Creates a vector model using the input corpus.

VectorModel(ent::EntropyWeighting, lw::LocalWeighting, corpus::BOW, labels;
    mindocs::Integer=1,
    smooth::Float64=0.0,
    weights=:balance,
    lowerweight=0.0
)

Creates a vector model using the input corpus.

Missing docstring.

Missing docstring for vectorize(::VectorModel, a, b). Check Documenter's build log for details.

Missing docstring.

Missing docstring for prune(::VectorModel, a, b). Check Documenter's build log for details.

Missing docstring.

Missing docstring for prune_select_top(::VectorModel, f). Check Documenter's build log for details.

Weighting methods for VectorModel

Local

TextSearch.TfWeighting — Type

TfWeighting()

Term frequency weighting

TextSearch.TpWeighting — Type

TpWeighting()

Term probability weighting

Missing docstring.

Missing docstring for FreqWeighthing. Check Documenter's build log for details.

TextSearch.BinaryLocalWeighting — Type

BinaryLocalWeighting()

The weight is 1 for known tokens, 0 for out of vocabulary tokens

Global

TextSearch.IdfWeighting — Type

IdfWeighting()

Inverse document frequency weighting

TextSearch.EntropyWeighting — Type

EntropyWeighting(; smooth=0.0, lowerweight=0.0, weights=:balance)

Entropy weighting uses the empirical entropy of the vocabulary along classes to produce a notion of importance for each token

TextSearch.BinaryGlobalWeighting — Type

BinaryGlobalWeighting()

The weight is 1 for known tokens, 0 for out of vocabulary tokens