Sparse matrices

Inverted indexes/files are representations of sparse matrices optimized for certain operations. We provide some functions to convert inverted files to sparse matrices.

Missing docstring.

Missing docstring for sparse. Check Documenter's build log for details.

SparseArrays.sparsevec — Function

sparsevec(vec::DVEC{Ti,Tv}, m=0) where {Ti<:Integer,Tv<:Number}

Creates a sparse vector from a DVEC sparse vector

Inverted indexes constructors also support sparse matrices as input (wrapped on MatrixDatabase structs)

Dictionary-based sparse vectors

Some application domains could take advantage of hash based sparse vectors, like text search and classification. Therefore, this package also provide a partial implementation of sparse vectors using Dict.

TextSearch.dvec — Function

dvec(x::AbstractSparseVector)

Converts an sparse vector into a DVEC sparse vector

Missing docstring.

Missing docstring for DVEC. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SVEC. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SVEC32. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SVEC64. Check Documenter's build log for details.

Missing docstring.

Missing docstring for nnz. Check Documenter's build log for details.

Missing docstring.

Missing docstring for findmax. Check Documenter's build log for details.

Missing docstring.

Missing docstring for argmax. Check Documenter's build log for details.

Missing docstring.

Missing docstring for maximum. Check Documenter's build log for details.

Missing docstring.

Missing docstring for findmin. Check Documenter's build log for details.

Missing docstring.

Missing docstring for argmin. Check Documenter's build log for details.

Missing docstring.

Missing docstring for minimum. Check Documenter's build log for details.

LinearAlgebra.normalize! — Function

normalize!(bow::DVEC)

Inplace normalization of bow

LinearAlgebra.dot — Function

dot(a::DVEC, b::DVEC)::Float64

Computes the dot product for two DVEC vectors

LinearAlgebra.norm — Function

norm(a::DVEC)::Float64

Computes a normalized DVEC vector

Base.zero — Function

zero(::Type{DVEC{Ti,Tv}}) where {Ti,Tv}

Creates an empty DVEC vector

TextSearch.add! — Function

add!(a::DVEC{Ti,Tv}, b::DVEC{Ti,Tv}) where {Ti,Tv<:Real}
add!(a::DVEC{Ti,Tv}, b::AbstractSparseArray) where {Ti,Tv<:Real}
add!(a::DVEC{Ti,Tv}, b::Pair{Ti,Tv}) where {Ti,Tv<:Real}

Updates a to the sum of a+b

Base.sum — Function

Base.sum(col::AbstractVector{<:DVEC})

Computes the sum of the given list of vectors

Base.:+ — Function

+(a::DVEC{Ti,Tv}, b::DVEC{Ti,Tv}) where {Ti,Tv<:Real}
+(a::DVEC, b::Pair)

Computes the sum of a and b

Base.:- — Function

-(a::DVEC{Ti,Tv}, b::DVEC{Ti,Tv}) where {Ti,Tv<:Real}

Substracts of b of a

Base.:* — Function

*(a::DVEC{Ti,Tv}, b::DVEC{Ti,Tv}) where {Ti,Tv<:Real}
*(a::DVEC{K, V}, b::F) where K where {V<:Real} where {F<:Real}

Computes the element-wise product of a and b

Base.:/ — Function

/(a::DVEC{K, V}, b::F) where K where {V<:Real} where {F<:Real}

Computes the element-wise division of a and b

TextSearch.centroid — Function

centroid(cluster::AbstractVector{<:DVEC})

Computes a centroid of the given list of DVEC vectors

Distances.evaluate — Function

evaluate(::NormalizedCosineDistance, a::DVEC, b::DVEC)::Float64

Computes the cosine distance between two DVEC sparse vectors

It supposes that bags are normalized (see normalize! function)

evaluate(::CosineDistance, a::DVEC, b::DVEC)::Float64

Computes the cosine distance between two DVEC sparse vectors

evaluate(::NormalizedAngleDistance, a::DVEC, b::DVEC)::Float64

Computes the angle between two DVEC sparse vectors

It supposes that all bags are normalized (see normalize! function)

evaluate(::AngleDistance, a::DVEC, b::DVEC)::Float64

Computes the angle between two DVEC sparse vectors

Missing docstring.

Missing docstring for NormalizedAngleDistance. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NormalizedCosineDistance. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AngleDistance. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CosineDistance. Check Documenter's build log for details.

Missing docstring.

Missing docstring for evaluate. Check Documenter's build log for details.