TextSearch API

TextSearch.bow — Method

bow(model::Tokenizer, x::AbstractSparseVector)
bow(model::Tokenizer, x::DVEC{Ti,Tv}) where {Ti<:Integer,Tv<:Number}

Creates a bag of words using the sparse vector x and the text model model

TextSearch.compute_bow — Function

compute_bow(messages::AbstractVector, bow::BOW=BOW())

Computes a bag of words from messages

TextSearch.compute_bow — Function

compute_bow(tokenlist::AbstractVector{S}, bow::BOW=BOW()) where {S<:Symbol}
compute_bow(tok::Tokenizer, text::Symbol, bow::BOW=BOW())
compute_bow(tok::Tokenizer, text::AbstractVector, bow::BOW=BOW())

Creates a bag of words from the given text (a string or a list of strings). If bow is given then updates the bag with the text. When config is given, the text is parsed according to it.

TextSearch.compute_bow_corpus — Function

compute_bow_corpus(tok::Tokenizer, corpus::AbstractVector)

Computes a list of bag of words from a corpus

TextSearch.dvec — Method

dvec(m::Tokenizer, x::DVEC{Symbol,Tv}, Ti=Int) where {Tv<:Number}

Creates a DVEC sparse vector from a bag of words sparse vector (i.e., with type DVEC{Symbol,Tv}), using the inverse map m

TextSearch.flush_token! — Method

flush_token!(tok::Tokenizer)

Pushes the word inside buffer into token list; it discards empty strings.

TextSearch.normalize_text — Method

normalize_text(config::TextConfig, text::AbstractString, output::Vector{Char})

Normalizes a given text using the specified transformations of config

TextSearch.nwords — Method

nwords(tok::Tokenizer, q::Integer)

TextSearch.prune — Method

prune(model::VectorModel, lowerweight::AbstractFloat)

Creates a new vector model without terms with smaller global weights than lowerweight.

TextSearch.prune_select_top — Method

prune_select_top(model::VectorModel, k::Integer)
prune_select_top(model::VectorModel, ratio::AbstractFloat)

Creates a new vector model with the best k tokens from model based on global weighting. ratio is a floating point between 0 and 1 indicating the ratio of the vocabulary to be kept

TextSearch.qgrams — Method

qgrams(tok::Tokenizer, q::Integer)

Computes character q-grams for the given input

TextSearch.skipgrams — Method

skipgrams(tok::Tokenizer, q::Skipgram)

Tokenizes using skipgrams

TextSearch.tokenize — Method

tokenize(tok::Tokenizer, text::AbstractString)

Tokenizes text using the given configuration

TextSearch.tokenize_corpus — Method

tokenize_corpus(tok::Tokenizer, arr)

Tokenize a list of texts.

TextSearch.unigrams — Method

unigrams(tok::Tokenizer, dropunigrams::Bool)

Performs the word tokenization

TextSearch.vectorize — Method

vectorize(model::VectorModel, bow::BOW, maxfreq::Integer=maximum(bow); normalize=true) where Tv<:Real

Computes a weighted vector using the given bag of words and the specified weighting scheme.

TextSearch.BinaryGlobalWeighting — Type

BinaryGlobalWeighting()

The weight is 1 for known tokens, 0 for out of vocabulary tokens

TextSearch.BinaryLocalWeighting — Type

BinaryLocalWeighting()

The weight is 1 for known tokens, 0 for out of vocabulary tokens

TextSearch.EntropyWeighting — Type

EntropyWeighting(; smooth=0.0, lowerweight=0.0, weights=:balance)

Entropy weighting uses the empirical entropy of the vocabulary along classes to produce a notion of importance for each token

TextSearch.FreqWeighting — Type

FreqWeighting()

Frequency weighting

TextSearch.GlobalWeighting — Type

GlobalWeighting

Abstract type for global weighting

TextSearch.IdfWeighting — Type

IdfWeighting()

Inverse document frequency weighting

TextSearch.LocalWeighting — Type

LocalWeighting

Abstract type for local weighting

TextSearch.Skipgram — Type

Skipgram(qsize, skip)

A skipgram is a kind of tokenization where qsize words having skip separation are used as a single token.

TextSearch.TextConfig — Type

TextConfig(;
    del_diac::Bool=true,
    del_dup::Bool=false,
    del_punc::Bool=false,
    group_num::Bool=true,
    group_url::Bool=true,
    group_usr::Bool=false,
    group_emo::Bool=false,
    lc::Bool=true,
    qlist::Vector=Int8[],
    nlist::Vector=Int8[],
    slist::Vector{Skipgram}=Skipgram[]
)

Defines a preprocessing and tokenization pipeline

del_diac: indicates if diacritic symbols should be removed

del_dup: indicates if duplicate contiguous symbols must be replaced for a single symbol

del_punc: indicates if punctuaction symbols must be removed

group_num: indicates if numbers should be grouped _num

group_url: indicates if urls should be grouped as _url

group_usr: indicates if users (@usr) should be grouped as _usr

group_emo: indicates if emojis should be grouped as _emo

lc: indicates if the text should be normalized to lower case

qlist: a list of character q-grams to use

nlist: a list of words n-grams to use

slist: a list of skip-grams tokenizers to use

Note: If qlist, nlist, and slists are all empty arrays, then it defaults to nlist=[1]

TextSearch.TextModel — Type

Model

An abstract type that represents a weighting model

TextSearch.TfWeighting — Type

TfWeighting()

Term frequency weighting

TextSearch.TokenStats — Type

TokenStats(occs, ndocs)

Stores useful information for a token; i.e., the number of occurrences in the collection, the number of documents having that token

TextSearch.Tokenizer — Type

struct Tokenizer

A tokenizer converts a text into a set of tokens, and in particular, each token in this implementation is represented as the hash code of the corresponding string token. This methods also normalize and preprocess the text following instructions in the given TextConfig object. The structure has several fields:

the text config object

an inverse map invmap that need to be captured at construction time (only needed if the tokens need to be inspected).

isconstruction indicator (must be true only when the tokens are being parsed for building a model)

the rest of the fields are used as buffers (multithreaded applications need independent copies of tokenizers)

Note: non-thread safe, make a copy of this structure for each thread.

TextSearch.TpWeighting — Type

TpWeighting()

Term probability weighting

TextSearch.VectorModel — Type

VectorModel(ent::EntropyWeighting, lw::LocalWeighting, corpus::BOW, labels;
    mindocs::Integer=1,
    smooth::Float64=0.0,
    weights=:balance,
    lowerweight=0.0
)

Creates a vector model using the input corpus.

TextSearch.VectorModel — Method

VectorModel(global_weighting::GlobalWeighting, local_weighting::LocalWeighting, corpus; mindocs=1)

Creates a vector model using the input corpus.