TextSearch API

Base.:* — Method

*(a::DVEC{Ti,Tv}, b::DVEC{Ti,Tv}) where {Ti,Tv<:Real}
*(a::DVEC{K, V}, b::F) where K where {V<:Real} where {F<:Real}

Computes the element-wise product of a and b

Base.:+ — Method

+(a::DVEC{Ti,Tv}, b::DVEC{Ti,Tv}) where {Ti,Tv<:Real}
+(a::DVEC, b::Pair)

Computes the sum of a and b

Base.:- — Method

-(a::DVEC{Ti,Tv}, b::DVEC{Ti,Tv}) where {Ti,Tv<:Real}

Substracts of b of a

Base.:/ — Method

/(a::DVEC{K, V}, b::F) where K where {V<:Real} where {F<:Real}

Computes the element-wise division of a and b

Base.append! — Method

append!(index::InvIndex, vectors)

Insert all elements vectors into index; each element (vector) is identified by its position in vectors

Base.maximum — Method

Base.maximum(voc::DVEC)

Computes the maximum (weight) value in a BOW

Base.push! — Method

push!(index::InvIndex, p::Pair{Int,SVEC})

Inserts a weighted vector into the index.

Base.sum — Method

Base.sum(col::AbstractVector{T}) where {T<:DVEC}

Computes the sum of the given list of vectors

Base.zero — Method

zero(::Type{DVEC{Ti,Tv}}) where {Ti,Tv}

Creates an empty DVEC vector

Distances.evaluate — Method

evaluate(::NormalizedAngleDistance, a::DVEC, b::DVEC)::Float64

Computes the angle between two DVEC sparse vectors

It supposes that all bags are normalized (see normalize! function)

Distances.evaluate — Method

evaluate(::NormalizedCosineDistance, a::DVEC, b::DVEC)::Float64

Computes the cosine distance between two DVEC sparse vectors

It supposes that bags are normalized (see normalize! function)

Distances.evaluate — Method

evaluate(::CosineDistance, a::DVEC, b::DVEC)::Float64

Computes the cosine distance between two DVEC sparse vectors

LinearAlgebra.dot — Method

dot(a::DVEC, b::DVEC)::Float64

Computes the dot product for two DVEC vectors

LinearAlgebra.norm — Method

norm(a::DVEC)::Float64

Computes a normalized DVEC vector

LinearAlgebra.normalize! — Method

normalize!(bow::DVEC)

Inplace normalization of bow

SparseArrays.sparse — Method

sparse(cols::AbstractVector{S}, m=0) where S<:DVEC{Ti,Tv} where {Ti<:Integer,Tv<:Number}

Creates a sparse matrix from an array of DVEC sparse vectors.

SparseArrays.sparsevec — Method

sparsevec(vec::DVEC{Ti,Tv}, m=0) where {Ti<:Integer,Tv<:Number}

Creates a sparse vector from a´ DVEC sparse vector

TextSearch._svs — Method

_svs(T::Type, sets::AbstractVector, by::Function)

Computes the intersection of the ordered lists in sets using the by::Function to extract a comparable for elements in each list

TextSearch.add! — Method

add!(a::DVEC{Ti,Tv}, b::DVEC{Ti,Tv}) where {Ti,Tv<:Real}
add!(a::DVEC{Ti,Tv}, b::AbstractSparseArray) where {Ti,Tv<:Real}
add!(a::DVEC{Ti,Tv}, b::Pair{Ti,Tv}) where {Ti,Tv<:Real}

Updates a to the sum of a+b

TextSearch.baezayates — Method

baezayates(first::AbstractVector, byfirst::Function, second::AbstractVector{T}, bysecond::Function, output::AbstractVector{T}) where T

Computes the intersection between first and second ordered lists using the Baeza-Yates algorithm [cite]; elements are mapped to a comparable value using byfirst and bysecond functions, for first and second lists. The matched objects of second are stored in output.

TextSearch.bow — Method

bow(model::Tokenizer, x::AbstractSparseVector)
bow(model::Tokenizer, x::DVEC{Ti,Tv}) where {Ti<:Integer,Tv<:Number}

Creates a bag of words using the sparse vector x and the text model model

TextSearch.centroid — Method

centroid(cluster::AbstractVector{<:DVEC})

Computes a centroid of the given list of DVEC vectors

TextSearch.compute_bow — Function

compute_bow(tokenlist::AbstractVector{S}, bow::BOW=BOW()) where {S<:Symbol}
compute_bow(tok::Tokenizer, text::Symbol, bow::BOW=BOW())
compute_bow(tok::Tokenizer, text::AbstractVector, bow::BOW=BOW())

Creates a bag of words from the given text (a string or a list of strings). If bow is given then updates the bag with the text. When config is given, the text is parsed according to it.

TextSearch.compute_bow — Function

compute_bow(messages::AbstractVector, bow::BOW=BOW())

Computes a bag of words from messages

TextSearch.compute_bow_corpus — Function

compute_bow_corpus(tok::Tokenizer, corpus::AbstractVector)

Computes a list of bag of words from a corpus

TextSearch.dvec — Method

dvec(x::AbstractSparseVector)

Converts an sparse vector into a DVEC sparse vector

TextSearch.dvec — Method

dvec(m::Tokenizer, x::DVEC{Symbol,Tv}, Ti=Int) where Tv<:Number

Creates a DVEC sparse vector from a bag of words sparse vector (i.e., with type DVEC{Symbol,Tv}), using the inverse map m

TextSearch.find_insert_position — Method

findinsertposition(arr::AbstractVector, value)

Finds the insert position of value inside arr. Note: it returns 0 for lower limit.

TextSearch.flush_token! — Method

flush_token!(tok::Tokenizer)

Pushes the word inside buffer into token list; it discards empty strings.

TextSearch.intersection — Method

intersection(sets::AbstractVector{S}, by::Function=identity) where {S<:AbstractVector}

Computes the intersection of sets represented by ordered arrays lists using the by::Function to extract a comparable for elements in each list

TextSearch.normalize_text — Method

normalize_text(config::TextConfig, text::AbstractString, output::Vector{Char})

Normalizes a given text using the specified transformations of config

TextSearch.nwords — Method

nwords(tok::Tokenizer, q::Integer)

TextSearch.prune — Method

prune(invindex::InvIndex, keeptop::Integer)
prune(invindex::InvIndex, keeptop_ratio::AbstractFloat)

Creates a new inverted index using the given invindex; the idea is to discard document entries with low weight. It keeps at most keeptop entries for each posting list.

TextSearch.prune — Method

prune(model::VectorModel, lowerweight::AbstractFloat)

Creates a new vector model without terms with smaller global weights than lowerweight.

TextSearch.prune_select_top — Method

prune_select_top(model::VectorModel, k::Integer)
prune_select_top(model::VectorModel, ratio::AbstractFloat)

Creates a new vector model with the best k tokens from model based on global weighting. ratio is a floating point between 0 and 1 indicating the ratio of the vocabulary to be kept

TextSearch.qgrams — Method

qgrams(tok::Tokenizer, q::Integer)

Computes character q-grams for the given input

TextSearch.skipgrams — Method

skipgrams(tok::Tokenizer, q::Skipgram)

Tokenizes using skipgrams

TextSearch.tokenize — Method

tokenize(tok::Tokenizer, text::AbstractString)

Tokenizes text using the given configuration

TextSearch.tokenize_corpus — Method

tokenize_corpus(tok::Tokenizer, arr)

Tokenize a list of texts.

TextSearch.unigrams — Method

unigrams(tok::Tokenizer, dropunigrams::Bool)

Performs the word tokenization

TextSearch.update! — Method

update!(a::InvIndex, b::InvIndex)

Updates inverted index a with b.

TextSearch.vectorize — Method

vectorize(model::VectorModel, bow::BOW, maxfreq::Integer=maximum(bow); normalize=true) where Tv<:Real

Computes a weighted vector using the given bag of words and the specified weighting scheme.

TextSearch.BinaryGlobalWeighting — Type

BinaryGlobalWeighting()

The weight is 1 for known tokens, 0 for out of vocabulary tokens

TextSearch.BinaryLocalWeighting — Type

BinaryLocalWeighting()

The weight is 1 for known tokens, 0 for out of vocabulary tokens

TextSearch.EntropyWeighting — Type

EntropyWeighting(; smooth=0.0, lowerweight=0.0, weights=:balance)

Entropy weighting uses the empirical entropy of the vocabulary along classes to produce a notion of importance for each token

TextSearch.FreqWeighting — Type

FreqWeighting()

Frequency weighting

TextSearch.GlobalWeighting — Type

GlobalWeighting

Abstract type for global weighting

TextSearch.IdfWeighting — Type

IdfWeighting()

Inverse document frequency weighting

TextSearch.LocalWeighting — Type

LocalWeighting

Abstract type for local weighting

TextSearch.Skipgram — Type

Skipgram(qsize, skip)

A skipgram is a kind of tokenization where qsize words having skip separation are used as a single token.

TextSearch.TextConfig — Type

TextConfig(;
    del_diac::Bool=true,
    del_dup::Bool=false,
    del_punc::Bool=false,
    group_num::Bool=true,
    group_url::Bool=true,
    group_usr::Bool=false,
    group_emo::Bool=false,
    lc::Bool=true,
    qlist::Vector=Int8[],
    nlist::Vector=Int8[],
    slist::Vector{Skipgram}=Skipgram[]
)

Defines a preprocessing and tokenization pipeline

del_diac: indicates if diacritic symbols should be removed

del_dup: indicates if duplicate contiguous symbols must be replaced for a single symbol

del_punc: indicates if punctuaction symbols must be removed

group_num: indicates if numbers should be grouped _num

group_url: indicates if urls should be grouped as _url

group_usr: indicates if users (@usr) should be grouped as _usr

group_emo: indicates if emojis should be grouped as _emo

lc: indicates if the text should be normalized to lower case

qlist: a list of character q-grams to use

nlist: a list of words n-grams to use

slist: a list of skip-grams tokenizers to use

Note: If qlist, nlist, and slists are all empty arrays, then it defaults to nlist=[1]

TextSearch.TextModel — Type

Model

An abstract type that represents a weighting model

TextSearch.TfWeighting — Type

TfWeighting()

Term frequency weighting

TextSearch.TokenStats — Type

TokenStats(occs, ndocs)

Stores useful information for a token; i.e., the number of occurrences in the collection, the number of documents having that token

TextSearch.Tokenizer — Type

struct Tokenizer

A tokenizer converts a text into a set of tokens, and in particular, each token in this implementation is represented as the hash code of the corresponding string token. This methods also normalize and preprocess the text following instructions in the given TextConfig object. The structure has several fields:

the text config object

an inverse map invmap that need to be captured at construction time (only needed if the tokens need to be inspected).

isconstruction indicator (must be true only when the tokens are being parsed for building a model)

the rest of the fields are used as buffers (multithreaded applications need independent copies of tokenizers)

Note: non-thread safe, make a copy of this structure for each thread.

TextSearch.TpWeighting — Type

TpWeighting()

Term probability weighting

TextSearch.VectorModel — Type

VectorModel(ent::EntropyWeighting, lw::LocalWeighting, corpus::BOW, labels;
    mindocs::Integer=1,
    smooth::Float64=0.0,
    weights=:balance,
    lowerweight=0.0
)

Creates a vector model using the input corpus.

TextSearch.VectorModel — Method

VectorModel(global_weighting::GlobalWeighting, local_weighting::LocalWeighting, corpus; mindocs=1)

Creates a vector model using the input corpus.