TextSearch API

Base.:* — Method

*(a::DVEC{Ti,Tv}, b::DVEC{Ti,Tv}) where {Ti,Tv<:Real}
*(a::DVEC{K, V}, b::F) where K where {V<:Real} where {F<:Real}

Computes the element-wise product of a and b

Base.:+ — Method

+(a::DVEC{Ti,Tv}, b::DVEC{Ti,Tv}) where {Ti,Tv<:Real}
+(a::DVEC, b::Pair)

Computes the sum of a and b

Base.:- — Method

-(a::DVEC{Ti,Tv}, b::DVEC{Ti,Tv}) where {Ti,Tv<:Real}

Substracts of b of a

Base.:/ — Method

/(a::DVEC{K, V}, b::F) where K where {V<:Real} where {F<:Real}

Computes the element-wise division of a and b

Base.sum — Method

Base.sum(col::AbstractVector{<:DVEC})

Computes the sum of the given list of vectors

Base.zero — Method

zero(::Type{DVEC{Ti,Tv}}) where {Ti,Tv}

Creates an empty DVEC vector

Distances.evaluate — Method

evaluate(::AngleDistance, a::DVEC, b::DVEC)::Float64

Computes the angle between two DVEC sparse vectors

Distances.evaluate — Method

evaluate(::CosineDistance, a::DVEC, b::DVEC)::Float64

Computes the cosine distance between two DVEC sparse vectors

Distances.evaluate — Method

evaluate(::NormalizedAngleDistance, a::DVEC, b::DVEC)::Float64

Computes the angle between two DVEC sparse vectors

It supposes that all bags are normalized (see normalize! function)

Distances.evaluate — Method

evaluate(::NormalizedCosineDistance, a::DVEC, b::DVEC)::Float64

Computes the cosine distance between two DVEC sparse vectors

It supposes that bags are normalized (see normalize! function)

LinearAlgebra.dot — Method

dot(a::DVEC, b::DVEC)::Float64

Computes the dot product for two DVEC vectors

LinearAlgebra.norm — Method

norm(a::DVEC)::Float64

Computes a normalized DVEC vector

LinearAlgebra.normalize! — Method

normalize!(bow::DVEC)

Inplace normalization of bow

SparseArrays.sparse — Method

sparse(cols::AbstractVector{S}, m=0; minweight=1e-9) where S<:DVEC{Ti,Tv} where {Ti<:Integer,Tv<:Number}

Creates a sparse matrix from an array of DVEC sparse vectors.

SparseArrays.sparsevec — Method

sparsevec(vec::DVEC{Ti,Tv}, m=0) where {Ti<:Integer,Tv<:Number}

Creates a sparse vector from a DVEC sparse vector

TextSearch.add! — Method

add!(a::DVEC{Ti,Tv}, b::DVEC{Ti,Tv}) where {Ti,Tv<:Real}
add!(a::DVEC{Ti,Tv}, b::AbstractSparseArray) where {Ti,Tv<:Real}
add!(a::DVEC{Ti,Tv}, b::Pair{Ti,Tv}) where {Ti,Tv<:Real}

Updates a to the sum of a+b

TextSearch.centroid — Method

centroid(cluster::AbstractVector{<:DVEC})

Computes a centroid of the given list of DVEC vectors

TextSearch.dvec — Method

dvec(x::AbstractSparseVector)

Converts an sparse vector into a DVEC sparse vector

TextSearch.filter_tokens — Method

filter_tokens(pred::Function, voc::Vocabulary)

Returns a copy of reduced vocabulary based on evaluating pred function for each entry in voc

TextSearch.flush_nword! — Method

flush_nword!(buff::TextSearchBuffer, tt::AbstractTokenTransformation)

Pushes the nword inside the buffer to the token list; it discards empty strings.

TextSearch.flush_qgram! — Method

flush_qgram!(buff::TextSearchBuffer, tt::AbstractTokenTransformation)

Pushes the qgram inside the buffer to the token list; it discards empty strings.

TextSearch.flush_skipgram! — Method

flush_skipgram!(buff::TextSearchBuffer, tt::AbstractTokenTransformation)

Pushes the skipgram inside the buffer to the token list; it discards empty strings.

TextSearch.flush_unigram! — Method

flush_unigram!(buff::TextSearchBuffer, tt::AbstractTokenTransformation)

Pushes the word inside the buffer to the token list; it discards empty strings.

TextSearch.normalize_text — Method

normalize_text(config::TextConfig, text::AbstractString, output::Vector{Char})

Normalizes a given text using the specified transformations of config

TextSearch.nwords — Method

nwords(q::Integer, buff::TextSearchBuffer, tt::AbstractTokenTransformation)

TextSearch.prune — Method

prune(model::VectorModel, lowerweight::AbstractFloat)

Creates a new vector model without terms with smaller global weights than lowerweight.

TextSearch.prune_select_top — Method

prune_select_top(model::VectorModel, k::Integer)
prune_select_top(model::VectorModel, ratio::AbstractFloat)

Creates a new vector model with the best k tokens from model based on global weighting. ratio is a floating point between 0 and 1 indicating the ratio of the vocabulary to be kept

TextSearch.qgrams — Method

qgrams(q::Integer, buff::TextSearchBuffer, tt::AbstractTokenTransformation)

Computes character q-grams for the given input

TextSearch.skipgrams — Method

skipgrams(q::Skipgram, buff::TextSearchBuffer, tt::AbstractTokenTransformation)

Tokenizes using skipgrams

TextSearch.tokenize — Function

tokenize(textconfig::TextConfig, text::AbstractString, buff=TextSearchBuffer(), makecopy=true)

Tokenizes text using the given configuration

TextSearch.tokenize_corpus — Method

tokenize_corpus(textconfig::TextConfig, arr; minbatch=0)

Tokenize a list of texts.

TextSearch.transform_nword — Method

transform_nword(::AbstractTokenTransformation, tok)

Hook applied in the tokenization stage to change the input token tok if needed. For instance, it can be used to apply stemming or any other kind of normalization. Return nothing to ignore the tok occurence (e.g., stop words).

TextSearch.transform_qgram — Method

transform_qgram(::AbstractTokenTransformation, tok)

Hook applied in the tokenization stage to change the input token tok if needed. For instance, it can be used to apply stemming or any other kind of normalization. Return nothing to ignore the tok occurence (e.g., stop words).

TextSearch.transform_skipgram — Method

transform_skipgram(::AbstractTokenTransformation, tok)

Hook applied in the tokenization stage to change the input token tok if needed. For instance, it can be used to apply stemming or any other kind of normalization. Return nothing to ignore the tok occurence (e.g., stop words).

TextSearch.transform_unigram — Method

transform_unigram(::AbstractTokenTransformation, tok)

Hook applied in the tokenization stage to change the input token tok if needed. For instance, it can be used to apply stemming or any other kind of normalization. Return nothing to ignore the tok occurence (e.g., stop words).

TextSearch.unigrams — Method

unigrams(buff::TextSearchBuffer, tt::AbstractTokenTransformation)

Performs the word tokenization

TextSearch.vectorize — Function

vectorize(voc::Vocabulary, tokenlist::AbstractVector, buff=BOW())
vectorize(voc::Vocabulary, textconfig::TextConfig, text::AbstractString, buff=BOW())

Creates a bag of words from the given text (a string or a list of strings). If bow is given then updates the bag with the text. When config is given, the text is parsed according to it.

TextSearch.vectorize — Method

vectorize(voc::Vocabulary, textconfig::TextConfig, messages::AbstractVector)

Computes a bag of words from messages

TextSearch.vectorize — Method

vectorize(model::VectorModel, bow::BOW; normalize=true, mindocs=model.mindocs, minweight=1e-9) where Tv<:Real

Computes a weighted vector using the given bag of words and the specified weighting scheme.

TextSearch.vectorize_corpus — Method

vectorize_corpus(textconfig::TextConfig, textconfig::TextConfig, corpus::AbstractVector; minbatch=0)

Computes a list of bag of words from a corpus

TextSearch.BinaryGlobalWeighting — Type

BinaryGlobalWeighting()

The weight is 1 for known tokens, 0 for out of vocabulary tokens

TextSearch.BinaryLocalWeighting — Type

BinaryLocalWeighting()

The weight is 1 for known tokens, 0 for out of vocabulary tokens

TextSearch.EntropyWeighting — Type

EntropyWeighting(; smooth=0.0, lowerweight=0.0, weights=:balance)

Entropy weighting uses the empirical entropy of the vocabulary along classes to produce a notion of importance for each token

TextSearch.FreqWeighting — Type

FreqWeighting()

Frequency weighting

TextSearch.GlobalWeighting — Type

GlobalWeighting

Abstract type for global weighting

TextSearch.IdfWeighting — Type

IdfWeighting()

Inverse document frequency weighting

TextSearch.LocalWeighting — Type

LocalWeighting

Abstract type for local weighting

TextSearch.Skipgram — Type

Skipgram(qsize, skip)

A skipgram is a kind of tokenization where qsize words having skip separation are used as a single token.

TextSearch.TextConfig — Type

TextConfig(;
    del_diac::Bool=true,
    del_dup::Bool=false,
    del_punc::Bool=false,
    group_num::Bool=true,
    group_url::Bool=true,
    group_usr::Bool=false,
    group_emo::Bool=false,
    lc::Bool=true,
    qlist::Vector=Int8[],
    nlist::Vector=Int8[],
    slist::Vector{Skipgram}=Skipgram[],
    tt=IdentityTokenTransformation()
)

Defines a preprocessing and tokenization pipeline

del_diac: indicates if diacritic symbols should be removed

del_dup: indicates if duplicate contiguous symbols must be replaced for a single symbol

del_punc: indicates if punctuaction symbols must be removed

group_num: indicates if numbers should be grouped _num

group_url: indicates if urls should be grouped as _url

group_usr: indicates if users (@usr) should be grouped as _usr

group_emo: indicates if emojis should be grouped as _emo

lc: indicates if the text should be normalized to lower case

qlist: a list of character q-grams to use

nlist: a list of words n-grams to use

slist: a list of skip-grams tokenizers to use

tt: An AbstractTokenTransformation struct

Note: If qlist, nlist, and slists are all empty arrays, then it defaults to nlist=[1]

TextSearch.TextModel — Type

Model

An abstract type that represents a weighting model

TextSearch.TfWeighting — Type

TfWeighting()

Term frequency weighting

TextSearch.TpWeighting — Type

TpWeighting()

Term probability weighting

TextSearch.VectorModel — Method

VectorModel(ent::EntropyWeighting, lw::LocalWeighting, corpus::BOW, labels;
    mindocs::Integer=1,
    smooth::Float64=0.0,
    weights=:balance,
    lowerweight=0.0
)

Creates a vector model using the input corpus.

TextSearch.VectorModel — Method

VectorModel(global_weighting::GlobalWeighting, local_weighting::LocalWeighting, voc::Vocabulary; mindocs=1)

Creates a vector model using the input corpus.

TextSearch.Vocabulary — Method

Vocabulary(corpus)

Computes a vocabulary from an already tokenized corpus

TextSearch.Vocabulary — Method

Vocabulary(n::Integer)

Creates a Vocabulary struct

TextSearch.Vocabulary — Method

Vocabulary(textconfig, corpus; minbatch=0)

Computes a vocabulary from a corpus using the TextConfig textconfig