VectorModel

TextSearch.IdFreq — Type

IdFreq(id, freq)

Stores a document identifier and its frequency

TextSearch.VectorModel — Type

VectorModel(weighting::WeightingType, corpus::BOW; minocc::Integer=1)

Trains a vector model using the input corpus.

Missing docstring.

Missing docstring for vectorize(::VectorModel, a, b). Check Documenter's build log for details.

Missing docstring.

Missing docstring for prune(::VectorModel, a, b). Check Documenter's build log for details.

Missing docstring.

Missing docstring for prune_select_top(::VectorModel, f). Check Documenter's build log for details.

Weighting methods for VectorModel

TextSearch.TfWeighting — Type

TfWeighting()

Term frequency weighting

TextSearch.TpWeighting — Type

TpWeighting()

Term probability weighting

TextSearch.IdfWeighting — Type

IdfWeighting()

Inverse document frequency weighting

TextSearch.TfidfWeighting — Type

TfidfWeighting()

TFIDF weighting

Missing docstring.

Missing docstring for FreqWeighthing. Check Documenter's build log for details.

EntModel

TextSearch.EntModel — Type

EntModel(model::DistModel, weighting::WeightingType; lower=0.0)
EntModel(weigthing::WeightingType, corpus, y; smooth=3, minocc=3, weights=:balance, lower=0.0, nclasses=0)

Fits an EntModel using the already fitted DistModel. It accepts only symbols with a final weight higher or equal than lower. Parameters: - corpus is text collection - y the set of associated labels (one-to-one with corpus) - smooth is a smoothing factor for the histogram. - weights accepts a list of weights (one per class) to be applied to the histogram - lower controls the minimum weight to be accepted - nclasses specifies the number of classes - minocc: minimum population to consider a token (without considering the smoothing factor).

Missing docstring.

Missing docstring for vectorize(::EntModel, a, b). Check Documenter's build log for details.

Missing docstring.

Missing docstring for prune(::EntModel, a, b). Check Documenter's build log for details.

Missing docstring.

Missing docstring for prune_select_top(::EntModel, f). Check Documenter's build log for details.

Weighting methods for EntModel

Missing docstring.

Missing docstring for EntTfWeighting. Check Documenter's build log for details.

Missing docstring.

Missing docstring for EntTpWeighting. Check Documenter's build log for details.

Missing docstring.

Missing docstring for EntWeighting. Check Documenter's build log for details.